refactor: create @hyperframes/studio-server package
