+++ packages/core/package.json
@@ -283,9 +283,9 @@
     "@chenglou/pretext": "^0.0.5",
     "@hyperframes/lint": "workspace:*",
     "@hyperframes/parsers": "workspace:*",
+    "@hyperframes/studio-server": "workspace:*",
     "bpm-detective": "^2.0.5",
-    "postcss": "^8.5.8",
-    "postcss-selector-parser": "^7.1.2"
+    "postcss": "^8.5.8"
   },
   "devDependencies": {
     "@types/jsdom": "^28.0.0",
@@ -295,14 +295,6 @@
     "tsx": "^4.21.0",
     "typescript": "^5.0.0",
     "vitest": "^3.2.4"
-  },
-  "peerDependencies": {
-    "hono": "^4.0.0"
-  },
-  "peerDependenciesMeta": {
-    "hono": {
-      "optional": true
-    }
   },
   "optionalDependencies": {
     "esbuild": "^0.25.12"
