--- packages/core/src/parsers/gsapUnroll.ts
@@ -1,2 +1,0 @@
-// ponytail: compat re-export — moved to @hyperframes/parsers
-export { unrollComputedTimeline } from "@hyperframes/parsers";
