+++ packages/core/src/studio-api/helpers/draftMarkers.ts
@@ -1,10 +1,2 @@
-/**
- * Draft-marker constants shared between core's PreviewAdapter and Studio's
- * manual-edits code. CSS custom properties written during a drag gesture, plus
- * the gesture marker attribute. Exported from @hyperframes/core/studio-api/draft-markers.
- */
-export const STUDIO_OFFSET_X_PROP = "--hf-studio-offset-x";
-export const STUDIO_OFFSET_Y_PROP = "--hf-studio-offset-y";
-export const STUDIO_WIDTH_PROP = "--hf-studio-width";
-export const STUDIO_HEIGHT_PROP = "--hf-studio-height";
-export const STUDIO_MANUAL_EDIT_GESTURE_ATTR = "data-hf-studio-manual-edit-gesture";
+/** @deprecated Import from @hyperframes/studio-server/draft-markers */
+export * from "@hyperframes/studio-server/draft-markers";
