+++ packages/core/src/studio-api/index.ts
@@ -1,18 +1,2 @@
-export { createStudioApi } from "./createStudioApi.js";
-export { createProjectSignature } from "./helpers/projectSignature.js";
-export type { StudioApiAdapter, ResolvedProject, RenderJobState, LintResult } from "./types.js";
-export { isSafePath, walkDir } from "./helpers/safePath.js";
-export { getMimeType, MIME_TYPES } from "./helpers/mime.js";
-export { buildSubCompositionHtml } from "./helpers/subComposition.js";
-export { getElementScreenshotClip, type ScreenshotClip } from "./helpers/screenshotClip.js";
-export {
-  STUDIO_MANUAL_EDITS_PATH,
-  createStudioManualEditsRenderBodyScript,
-  createStudioPositionSeekReapplyScript,
-  type StudioManualEditsRenderScriptOptions,
-} from "./helpers/manualEditsRenderScript.js";
-export {
-  STUDIO_MOTION_PATH,
-  createStudioMotionRenderBodyScript,
-  type StudioMotionRenderScriptOptions,
-} from "./helpers/studioMotionRenderScript.js";
+/** @deprecated Import from @hyperframes/studio-server */
+export * from "@hyperframes/studio-server";
