+++ packages/studio-server/src/helpers/mediaValidation.test.ts
