+++ packages/studio-server/src/helpers/mediaValidation.ts
