+++ packages/studio-server/src/helpers/mime.ts
