+++ packages/studio-server/src/helpers/previewAdapter.test.ts
