+++ packages/studio-server/src/helpers/previewAdapter.ts
