+++ packages/studio-server/src/helpers/projectSignature.ts
