+++ packages/studio-server/src/helpers/safePath.test.ts
