+++ packages/studio-server/src/helpers/waveform.ts
