+++ packages/studio-server/src/routes/registry.ts
