+++ packages/studio-server/src/routes/storyboard.ts
@@ -7,7 +7,7 @@ import {
   SCRIPT_FILENAME,
   STORYBOARD_FILENAME,
   type StoryboardFrame,
-} from "../../storyboard/index.js";
+} from "@hyperframes/core/storyboard";
 
 /** A frame enriched with disk-resolution info the Studio needs to render tiles. */
 interface ResolvedStoryboardFrame extends StoryboardFrame {
