+++ packages/studio-server/src/routes/thumbnail.test.ts
