+++ packages/studio-server/src/routes/thumbnail.ts
