+++ packages/studio-server/src/routes/waveform.ts
